+++ emails/templates/due-tasks.js
@@ -6,7 +6,7 @@ const {
   style,
   vspace,
 } = require('../template-utils')
-const { accountData } = require('../mock-utils')
+const { accountData } = require('../helpers')
 
 module.exports = function() {
   const statusOpenDue = `
