+++ emails/templates/generic-notification.js
@@ -7,7 +7,7 @@ const {
   table,
   trimMultiline,
 } = require('../template-utils')
-const { accountData } = require('../mock-utils')
+const { accountData } = require('../helpers')
 
 module.exports = function() {
   return {
