+++ emails/templates/ruled.js
@@ -5,7 +5,7 @@ const {
   infobox,
   link,
 } = require('../template-utils')
-const { accountData } = require('../mock-utils')
+const { accountData } = require('../helpers')
 
 module.exports = function() {
   return {
